Add onMessage(id, data) to plugin API.
A solution was needed to notify the audio player to pause when a phone call comes in.  The option was to add a specific onPhone() method or generalize it.  Since there are other "events" that are useful to plugins, a generalized solution was used.  It is also extensible without changing the plugin API again.

Any plugin can call DroidGap.onMessage() to send a message/event to all other plugins.  NetworkManager was updated to send changes in connection status to plugins, so they can intelligently handle lost connections.